+++ src/interfaces/response.ts
@@ -1,5 +1,0 @@
-export interface ApiResponse {
-  slug: string;
-  utility: string;
-  property: string[];
-}