+++ src/interfaces/utility.ts
@@ -1,0 +1,10 @@
+export interface Utility {
+  prefix: string;
+  properties: string[];
+  slug: string;
+  values: { [key: string]: string };
+}
+
+export interface Utilities {
+  [key: string]: Utility;
+}